스케줄러 위치 수정
