--- 6week/server-java/src/main/java/kr/hhplus/be/server/interfaces/api/order/scheduler/OrderScheduler.java
@@ -1,33 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.order.scheduler;
-
-import jakarta.transaction.Transactional;
-import kr.hhplus.be.server.application.order.OrderFacade;
-import kr.hhplus.be.server.domain.order.service.OrderService;
-import kr.hhplus.be.server.domain.order.vo.OrderHistoryProductGroupVo;
-import kr.hhplus.be.server.domain.product.model.DomainProduct;
-import kr.hhplus.be.server.infrastructure.product.entity.Product;
-import kr.hhplus.be.server.infrastructure.product.entity.ProductRank;
-import kr.hhplus.be.server.domain.product.service.ProductRankService;
-import kr.hhplus.be.server.domain.product.service.ProductService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-
-import java.util.LinkedList;
-import java.util.List;
-
-@Component
-@RequiredArgsConstructor
-@Slf4j
-public class OrderScheduler {
-
-    private final OrderFacade facade;
-
-
-    @Scheduled(cron = "0 0 0 * * *")
-    @Transactional
-    public void makeRank(){
-        facade.updateRank();
-    }
-}
